Modify Ui class diagrams
