+++ src/main/java/team/serenity/ui/lessondata/LessonDataPanel.java
@@ -14,7 +14,7 @@ import team.serenity.model.group.question.Question;
 import team.serenity.model.group.studentinfo.StudentInfo;
 import team.serenity.ui.DataPanel;
 import team.serenity.ui.groupdata.GroupDataPanel;
-import team.serenity.ui.serenitydata.QuestionListViewCell;
+import team.serenity.ui.serenitydata.SerenityDataPanel.QuestionListViewCell;
 
 public class LessonDataPanel extends DataPanel {
     private static final String FXML = "LessonDataPanel.fxml";
